feat: refactor helpers into being multiple classes
