+++ src/Dan.Plugin.Tilda/Models/TildaDataSource.cs
@@ -8,8 +8,10 @@ using CloudNative.CloudEvents.NewtonsoftJson;
 using Dan.Common.Models;
 using Dan.Plugin.Tilda.Config;
 using Dan.Plugin.Tilda.Exceptions;
+using Dan.Plugin.Tilda.Extensions;
 using Dan.Plugin.Tilda.Interfaces;
 using Dan.Plugin.Tilda.Models.AlertMessages;
+using Dan.Plugin.Tilda.Services;
 using Dan.Plugin.Tilda.Utils;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
@@ -47,12 +49,14 @@ namespace Dan.Plugin.Tilda.Models
         protected ILogger _logger;
         protected HttpClient _client;
         protected HttpClient _alertClient;
+        protected IUriFormatter _uriFormatter;
 
         public TildaDataSource(
             IOptions<Settings> settings,
             IHttpClientFactory httpClientFactory,
             ILoggerFactory loggerFactory,
-            ResiliencePipelineProvider<string> pipelineProvider)
+            ResiliencePipelineProvider<string> pipelineProvider,
+            IUriFormatter uriFormatter)
         {
             _pipelineProvider = pipelineProvider;
             _settings = settings.Value;
@@ -60,6 +64,7 @@ namespace Dan.Plugin.Tilda.Models
             _client = httpClientFactory.CreateClient("SafeHttpClient");
             _alertClient = httpClientFactory.CreateClient("AlertHttpClient");
             BaseUri = _settings.GetClassBaseUri(GetType().Name);
+            _uriFormatter = uriFormatter;
         }
 
         public TildaDataSource()
@@ -70,71 +75,72 @@ namespace Dan.Plugin.Tilda.Models
         {
             var url = GetUri(BaseUri, AuditReportDatasetName, req.OrganizationNumber, req.Requestor, fromDate,
                 toDate);
-            return await Helpers.GetData<AuditReportList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+
+            return await _client.GetData<AuditReportList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<AuditReportList> GetAuditReportsAllAsync(EvidenceHarvesterRequest req, string month, string year, string filter)
         {
             var url = GetUriAll(BaseUri, AuditReportDatasetName, req.Requestor, month, year, string.Empty, string.Empty, filter);
-            return await Helpers.GetData<AuditReportList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<AuditReportList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<TrendReportList> GetDataTrendAsync(EvidenceHarvesterRequest req, DateTime? fromDate, DateTime? toDate)
         {
             var url = GetUri(BaseUri, TrendDatasetName, req.OrganizationNumber, req.Requestor, fromDate, toDate);
-            return await Helpers.GetData<TrendReportList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<TrendReportList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<TrendReportList> GetDataTrendAllAsync(EvidenceHarvesterRequest req, string month, string year, string filter)
         {
             var url = GetUriAll(BaseUri, TrendDatasetName, req.Requestor, month, year, string.Empty, string.Empty, filter);
-            return await Helpers.GetData<TrendReportList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<TrendReportList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<AuditCoordinationList> GetAuditCoordinationAsync(EvidenceHarvesterRequest req, DateTime? fromDate, DateTime? toDate)
         {
             var url = GetUri(BaseUri, CoordinationDatasetName, req.OrganizationNumber, req.Requestor, fromDate, toDate);
-            return await Helpers.GetData<AuditCoordinationList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<AuditCoordinationList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<AuditCoordinationList> GetAuditCoordinationAllAsync(EvidenceHarvesterRequest req, string month, string year, string filter)
         {
             var url = GetUriAll(BaseUri, CoordinationDatasetName, req.Requestor, month, year, string.Empty, string.Empty, filter);
-            return await Helpers.GetData<AuditCoordinationList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<AuditCoordinationList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<NPDIDAuditReportList> GetNPDIDAuditReportsAsync(EvidenceHarvesterRequest req, DateTime? fromDate, DateTime? toDate, string npdid)
         {
             var url = GetUri(BaseUri, NpdidDatasetName, req.OrganizationNumber, req.Requestor, fromDate, toDate, null, npdid);
-            return await Helpers.GetData<NPDIDAuditReportList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<NPDIDAuditReportList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<NPDIDAuditReportList> GetNPDIDAuditReportsAllAsync(EvidenceHarvesterRequest req, string month, string year, string npdid, string filter)
         {
             var url = GetUriAll(BaseUri, NpdidDatasetName, req.Requestor, month, year, null, npdid);
-            return await Helpers.GetData<NPDIDAuditReportList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<NPDIDAuditReportList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual async Task<AlertMessageList> GetAlertMessagesAsync(EvidenceHarvesterRequest req, string month, string year, string identifier)
         {
             var url = GetUriAll(BaseUri, AlertDatasetName, req.Requestor, month, year, identifier, null);
-            return await Helpers.GetData<AlertMessageList>(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetData<AlertMessageList>(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual string GetUri(string baseUri, string dataset, string organizationNumber, string requestor, DateTime? fromDate, DateTime? toDate, string identifier = "", string npdid = "")
         {
-            return Helpers.GetUri(baseUri, dataset, organizationNumber, requestor, fromDate, toDate, null, npdid);
+            return _uriFormatter.GetUri(baseUri, dataset, organizationNumber, requestor, fromDate, toDate, null, npdid);
         }
 
         public virtual async Task<byte[]> GetPdfReport(EvidenceHarvesterRequest req, string internTilsynsId)
         {
             var url = GetUri(BaseUri, PdfReportDatasetName, req.Requestor, null, null, null, internTilsynsId);
-            return await Helpers.GetPdfreport(url, OrganizationNumber, _client, _logger, req.MPToken, req.Requestor);
+            return await _client.GetPdfreport(url, OrganizationNumber, _logger, req.MPToken, req.Requestor);
         }
 
         public virtual string GetUriAll(string baseUri, string dataset, string requestor, string month, string year, string identifier = "", string npdid = "", string filter = "")
         {
-            return Helpers.GetUriAll(baseUri, dataset, requestor, month, year, identifier, null, filter);
+            return _uriFormatter.GetUriAll(baseUri, dataset, requestor, month, year, identifier, null, filter);
         }
 
         public virtual async Task<List<AlertSourceMessage>> GetAlertMessagesAsync(string from)
