+++ src/Dan.Plugin.Tilda/TildaSources/Arbeidstilsynet.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using Dan.Plugin.Tilda.Config;
 using Dan.Plugin.Tilda.Interfaces;
 using Dan.Plugin.Tilda.Models;
+using Dan.Plugin.Tilda.Services;
 using Dan.Plugin.Tilda.Utils;
 using Microsoft.Extensions.Options;
 using Nadobe.Common.Models;
@@ -32,8 +33,9 @@ namespace Dan.Plugin.Tilda.TildaSources
         public Arbeidstilsynet(IOptions<Settings> settings,
             IHttpClientFactory httpClientFactory,
             ILoggerFactory loggerFactory,
-            ResiliencePipelineProvider<string> pipelineProvider) :
-            base(settings, httpClientFactory, loggerFactory, pipelineProvider)
+            ResiliencePipelineProvider<string> pipelineProvider,
+            IUriFormatter uriFormatter) :
+            base(settings, httpClientFactory, loggerFactory, pipelineProvider, uriFormatter)
         {
         }
