+++ src/Dan.Plugin.Tilda/TildaSources/Branntilsyn.cs
@@ -4,6 +4,8 @@ using System.Net.Http;
 using Dan.Plugin.Tilda.Config;
 using Dan.Plugin.Tilda.Interfaces;
 using Dan.Plugin.Tilda.Models;
+using Dan.Plugin.Tilda.Services;
+using Dan.Plugin.Tilda.Utils;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Polly.Registry;
@@ -25,8 +27,9 @@ namespace Dan.Plugin.Tilda.TildaSources
         public Branntilsyn(IOptions<Settings> settings,
             IHttpClientFactory httpClientFactory,
             ILoggerFactory loggerFactory,
-            ResiliencePipelineProvider<string> pipelineProvider) :
-            base(settings, httpClientFactory, loggerFactory, pipelineProvider)
+            ResiliencePipelineProvider<string> pipelineProvider,
+            IUriFormatter uriFormatter) :
+            base(settings, httpClientFactory, loggerFactory, pipelineProvider, uriFormatter)
         {
 
         }
