+++ src/Dan.Plugin.Tilda/TildaSources/NSO.cs
@@ -5,6 +5,8 @@ using System.Text;
 using Dan.Plugin.Tilda.Config;
 using Dan.Plugin.Tilda.Models;
 using Dan.Plugin.Tilda.Interfaces;
+using Dan.Plugin.Tilda.Services;
+using Dan.Plugin.Tilda.Utils;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Polly.Registry;
@@ -29,8 +31,9 @@ namespace Dan.Plugin.Tilda.TildaSources
         public NSO(IOptions<Settings> settings,
             IHttpClientFactory httpClientFactory,
             ILoggerFactory loggerFactory,
-            ResiliencePipelineProvider<string> pipelineProvider) :
-            base(settings, httpClientFactory, loggerFactory, pipelineProvider)
+            ResiliencePipelineProvider<string> pipelineProvider,
+            IUriFormatter uriFormatter) :
+            base(settings, httpClientFactory, loggerFactory, pipelineProvider, uriFormatter)
         {
 
         }
