+++ src/Dan.Plugin.Tilda/TildaSources/StatsforvalterenMoreOgRomsdal.cs
@@ -3,6 +3,8 @@ using System.Net.Http;
 using Dan.Plugin.Tilda.Config;
 using Dan.Plugin.Tilda.Interfaces;
 using Dan.Plugin.Tilda.Models;
+using Dan.Plugin.Tilda.Services;
+using Dan.Plugin.Tilda.Utils;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Polly.Registry;
@@ -21,8 +23,9 @@ namespace Dan.Plugin.Tilda.TildaSources
         public StatsforvalterenMoreOgRomsdal(IOptions<Settings> settings,
             IHttpClientFactory httpClientFactory,
             ILoggerFactory loggerFactory,
-            ResiliencePipelineProvider<string> pipelineProvider) :
-            base(settings, httpClientFactory, loggerFactory, pipelineProvider)
+            ResiliencePipelineProvider<string> pipelineProvider,
+            IUriFormatter uriFormatter) :
+            base(settings, httpClientFactory, loggerFactory, pipelineProvider, uriFormatter)
         {
 
         }
